requirements.txt

defusedxml==0.4.1
oauthlib==1.0.3
PyJWT==1.4.0
python-social-auth==0.2.13
python3-openid==3.0.9
requests==2.8.1
requests-oauthlib==0.5.0






coverage==4.0.2
coveralls==1.1
django-cors-headers==1.1.0
Django==1.8.6
django-bootstrap3==6.2.2
django-debug-toolbar==1.4
django-extensions==1.5.9
Pillow==3.0.0
pip==7.1.2
psycopg2==2.6.1
setuptools==18.2
six==1.10.0
sqlparse==0.1.18
wheel==0.24.0
djangorestframework==3.3.1
python-memcached==1.57
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
gunicorn==19.3.0
sendgrid-django
